requirements/requirements.txt
cython
plyfile

ujson  # make json loading in cocoapi_nvidia faster
pycocotools

cffi
ninja
black
docformatter
setproctitle
fastfunc
meshplex
OpenEXR
vispy>=0.6.4
yacs>=0.1.8
tabulate
pytest-runner
pytest
ipdb
tqdm
numba
mmcv-full
imagecorruptions
pyassimp==4.1.3  # 4.1.4 will cause egl_renderer SegmentFault
pypng
imgaug>=0.4.0
albumentations
transforms3d
# pyquaternion
torchvision
open3d
fvcore
tensorboardX
einops
pytorch3d
# timm  # pytorch-image-models
git+ssh://git@github.com/rwightman/pytorch-image-models.git  # the latest timm
glfw
imageio
imageio-ffmpeg
PyOpenGL  # >=3.1.5
PyOpenGL_accelerate
chardet

thop  # https://github.com/Lyken17/pytorch-OpCounter
loguru


pytorch-lightning  # tested for 1.6.0.dev0
fairscale
deepspeed

# verified versions
pyro-ppl==1.4.0 #for epro-pnp <1.8.1
onnx==1.8.1
onnxruntime==1.8.0
onnx-simplifier==0.3.5
